@@ -9,13 +9,21 @@ use Phpactor\Extension\LanguageServerPhpstan\Model\Linter;
 use Phpactor\Extension\LanguageServerPhpstan\Model\PhpstanProcess;
 use Phpactor\LanguageServerProtocol\Diagnostic;
 use Phpactor\TextDocument\TextDocumentUri;
+
 use function Safe\tempnam;
 use function Safe\file_put_contents;
 
 class PhpstanLinter implements Linter
 {
-    public function __construct(private PhpstanProcess $process)
+    public function __construct(
+        private PhpstanProcess $process,
+        private bool $disableTmpFile = false,
+    ) {
+    }
+
+    public function isTmpFileDisabled(): bool
     {
+        return $this->disableTmpFile;
     }
 
     public function lint(string $url, ?string $text): Promise
@@ -32,14 +40,17 @@ class PhpstanLinter implements Linter
      */
     private function doLint(string $url, ?string $text): Generator
     {
-        if (null === $text) {
+        if (null === $text || $this->disableTmpFile) {
             return yield $this->process->analyse(TextDocumentUri::fromString($url)->path());
         }
 
         $name = tempnam(sys_get_temp_dir(), 'phpstanls');
         file_put_contents($name, $text);
-        $diagnostics = yield $this->process->analyse($name);
-        unlink($name);
+        try {
+            $diagnostics = yield $this->process->analyse($name);
+        } finally {
+            @unlink($name);
+        }
         return $diagnostics;
     }
 }
