@@ -5,22 +5,26 @@ namespace Phpactor\Extension\Navigation\Application;
 use Phpactor\Extension\Navigation\Navigator\Navigator as NavigatorInterface;
 use RuntimeException;
 use Phpactor\Extension\CodeTransformExtra\Application\ClassNew;
+use Symfony\Component\Filesystem\Path;
 
 class Navigator
 {
+    /** @param array<string, string> $autoCreateConfig */
     public function __construct(
         private NavigatorInterface $navigator,
         private ClassNew $classNew,
-        private array $autoCreateConfig
+        private array $autoCreateConfig,
+        private string $absolutePath
     ) {
     }
 
-    public function destinationsFor(string $path)
+    /** @return array<string, string> */
+    public function destinationsFor(string $path): array
     {
         return $this->navigator->destinationsFor($path);
     }
 
-    public function canCreateNew(string $path, string $destinationName)
+    public function canCreateNew(string $path, string $destinationName): bool
     {
         $destination = $this->destination($path, $destinationName);
 
@@ -38,7 +42,7 @@ class Navigator
         $this->classNew->generate($destination, $variant);
     }
 
-    private function destination(string $path, string $destinationName)
+    private function destination(string $path, string $destinationName): string
     {
         $destinations = $this->destinationsFor($path);
 
@@ -50,10 +54,10 @@ class Navigator
             ));
         }
 
-        return $destinations[$destinationName];
+        return Path::makeAbsolute($destinations[$destinationName], $this->absolutePath);
     }
 
-    private function variant(string $destinationName)
+    private function variant(string $destinationName): string
     {
         if (!isset($this->autoCreateConfig[$destinationName])) {
             throw new RuntimeException(sprintf(
